+++ CHANGELOG.md
@@ -9,11 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## Added
 - README for the Docker file
 - Support for ARM and ARM64 docker images
+- Documented support for multiple debug classes
 
 ## Fixed
 - Web files again included in wheel
 
 ## Changed
+- Reduced logging for PGP Timestamping Server mail handling
 
 
 # 0.9.4 - 2020-05-09